% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Data preprocessing.}
\usage{
data_preprocessing(inputs)
}
\arguments{
\item{inputs}{Input a gene expression matrix normalized as CPM values as a data.frame object. Each row represent a sample and each column represent a gene.}
}
\value{
a preprocessed gene expression matrix for model prediction.
}
\description{
Data preprocessing.
}
