% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{model_predict}
\alias{model_predict}
\title{Subtyping brain tumor patients with different immune infiltration signatures based on expression profiles.}
\usage{
model_predict(data)
}
\arguments{
\item{An}{expression matrix of brain tumor transcriptomes. The gene expression should be normalized as CPM values. Please make sure each row of expression matrix represent a patient and each column represent a Hugo Symbol of human gene.}
}
\value{
A dataframe of subtyping result.
}
\description{
Subtyping brain tumor patients with different immune infiltration signatures based on expression profiles.
}
